+++ bothub_client/bot.py
@@ -2,6 +2,8 @@
 
 from __future__ import (absolute_import, division, print_function)
 
+import os
+
 from bothub_client.intent import IntentState
 from bothub_client.dispatcher import DefaultDispatcher
 
@@ -32,11 +34,13 @@ class BaseBot(object):
         :type event: dict
         :param context: a context Bot runs
         :type context: dict'''
-        content = event.get('content')
 
-        bot_dir_path = os.path.dirname(os.path.realpath(__file__))
-        yml_path = os.path.join(bot_dir_path, os.pardir, 'bothub.yml')
-        intent_slots = IntentState.load_intent_slots_from_yml(yml_path)
+        bot_dir_path = os.path.realpath('.')
+        yml_path = os.path.join(bot_dir_path, 'bothub.yml')
+        if os.path.isfile(yml_path):
+            intent_slots = IntentState.load_intent_slots_from_yml(yml_path)
+        else:
+            intent_slots = []
 
         state = IntentState(self, intent_slots)
         dispatcher = DefaultDispatcher(self, state)