+++ bothub_client/intent.py
@@ -1,16 +1,13 @@
 # -*- coding: utf-8 -*-
 
 import logging
-import json
 import yaml
 from collections import namedtuple
-
-from bothub_client.dispatcher import DefaultDispatcher
 
 logger = logging.getLogger('bothub.intent')
 
 Intent = namedtuple('Intent', ['id', 'on_complete', 'slots'])
-IntentResult = namedtuple('IntentResult', ['intent_id', 'completed', 'answers', 'next_message'])
+IntentResult = namedtuple('IntentResult', ['intent_id', 'completed', 'answers', 'next_message', 'complete_handler_name'])
 Slot = namedtuple('Slot', ['id', 'question', 'datatype'])
 
 
@@ -19,7 +16,10 @@ class NoSlotRemainsException(Exception):
 
 
 class IntentState(object):
-    '''Manage intent state'''
+    '''I manage intent state.
+
+    Store and manage intent state in Bot user data storage.
+    '''
     intent_id_field = '_intent_id'
     intent_answers_field = '_intent_answers'
     remaining_slots_field = '_remaining_slots'
@@ -31,16 +31,19 @@ class IntentState(object):
         self.intent_id_to_intent_definition = dict([(i.id, i) for i in intent_slots])
 
     @staticmethod
-    def load_intent_slots_from_yml(path):
+    def load_yml(path):
         with open(path) as fin:
             content = fin.read()
-        config = yaml.load(content)
+        return yaml.load(content)
 
+    @staticmethod
+    def load_intent_slots_from_yml(path):
+        config = IntentState.load_yml(path)
         intents = config.get('intents', {})
         intent_slots = []
         for intent_id in intents.keys():
             intent_yaml = intents[intent_id]
-            on_complete = intent_yaml.get('on_complete', 'on_{}'.format(intent_id))
+            on_complete = intent_yaml.get('on_complete', 'set_{}'.format(intent_id))
             slots_yaml = intent_yaml.get('slots', [])
             slot_objs = []
             for slot in slots_yaml:
@@ -53,7 +56,13 @@ class IntentState(object):
             intent_slots.append(intent)
         return intent_slots
 
-    def init(self, intent_id):
+    def open(self, intent_id):
+        '''Open and start an intent.
+        Should execute a `next()` method after open an intent.
+
+        :param intent_id: an intent ID
+        :type intent_id: str
+        '''
         logger.debug('IntentState: init %s', intent_id)
         data = self.bot.get_user_data()
         data[self.intent_id_field] = intent_id
@@ -64,36 +73,52 @@ class IntentState(object):
         ]
         self.bot.set_user_data(data)
 
-    def is_processing(self, data):
+    def is_opened(self):
+        '''Returns whether an intent is opened or not.
+
+        :rtype: bool
+        '''
+        data = self.bot.get_user_data()
         return self.intent_id_field in data and data[self.intent_id_field] is not None
 
     def next(self, event=None):
+        '''Proceed a current opened intent.
+        If a user answer is given, store the answer and return an IntentResult object.
+
+        :param event: an event dict
+        :type event: dict
+        :return: an intent result object
+        :rtype: IntentResult
+        '''
         logger.debug('IntentState: next with event %s', event)
         data = self.bot.get_user_data()
         if event:
             self._store_answer(event, data)
         result = self._make_result_obj(data)
-        if result.completed: self._clear_state(data)
+        if result.completed:
+            self._clear_state(data)
         self.bot.set_user_data(data)
         return result
 
-    def on_complete(self, intent_id, **kwargs):
+    def close(self):
+        '''Close an opened intent.
+
+        Clear intent state variables in user data storage.
+        '''
         data = self.bot.get_user_data()
-        logger.debug('IntentState: on_complete intent %s', intent_id)
-        intent = self.intent_id_to_intent_definition[intent_id]
-        func_name = intent.on_complete
-        func = getattr(self.bot, func_name)
-        func(**kwargs)
+        self._clear_state(data)
+        self.bot.set_user_data(data)
 
     def _make_result_obj(self, data):
         next_message = self._next_slot_message(data)
         completed = next_message is None
-        result = IntentResult(
-            data[self.intent_id_field],
-            completed,
-            dict(data[self.intent_answers_field].items()),
-            next_message
-        )
+        intent_id = data[self.intent_id_field]
+        intent = self.intent_id_to_intent_definition[intent_id]
+        result = IntentResult(intent_id,
+                              completed,
+                              dict(data[self.intent_answers_field].items()),
+                              next_message,
+                              intent.on_complete)
         return result
 
     def _clear_state(self, data):
@@ -108,16 +133,12 @@ class IntentState(object):
         slot_id = data[self.slot_id_field]
         data.setdefault(self.intent_answers_field, {})[slot_id] = event['content']
 
-    def _load_slot(self, data):
-        slot = data[self.remaining_slots_field].pop(0)
-
     def _has_remainig_slots(self, data):
-        return self.remaining_slots_field not in data or data[self.remaining_slots_field] is not None
-        
+        return self.remaining_slots_field in data and data[self.remaining_slots_field] is not None
+
     def _next_slot_message(self, data):
         try:
             if not self._has_remainig_slots(data):
-                self._clear_state(data)
                 self.bot.set_user_data(data)
                 raise NoSlotRemainsException()
             slot = data[self.remaining_slots_field].pop(0)