+++ CHANGELOG.rst
@@ -1,3 +1,16 @@
+0.1.24
+------
+
+* change: add `event`, `context` arguments to on_complete handler method
+* change: change a default on_complete handler method to have a prefix `set_` instead of `on_`
+
+0.1.23
+------
+
+* fix: add a missing import
+* fix: lookup another location to find bothub.yml file
+
+
 0.1.22
 ------